+++ biome.json
@@ -42,7 +42,8 @@
         "noUselessElse": "off",
         "noNonNullAssertion": "off",
         "noParameterAssign": "off",
-        "useTemplate": "off"
+        "useTemplate": "off",
+        "useImportType": "off"
       }
     }
   },
