+++ react/rangeslider/src/RangeSlider.tsx
@@ -1,5 +1,5 @@
 import { CoreSlider } from '../../slider/src/core/CoreSlider.tsx';
-import type { RangeSliderProps } from '../../slider/src/props.ts';
+import { RangeSliderProps } from './props.ts';
 
 export function RangeSlider(props: RangeSliderProps) {
   return CoreSlider(props);
