+++ react/slider/src/Slider.tsx
@@ -1,5 +1,5 @@
 import { CoreSlider } from './core/CoreSlider.tsx';
-import type { SliderProps } from './props.ts';
+import { SliderProps } from './props.ts';
 
 export function Slider(props: SliderProps) {
   return CoreSlider(props);
